Refactoring the project into modules.
