+++ Pykov/db_init.py
@@ -30,8 +30,8 @@ import hashlib
 import os
 import hashlib
 import re
-import Markov
 import pickle
+from . import Markov
 conn = sqlite3.connect('pykov.db')
 
 default_texts = []
@@ -43,12 +43,12 @@ c = conn.cursor()
 
 
 def init_corpus():
-	path = os.path.join("static", "texts")
+	path = os.path.join("Pykov", "static", "texts")
 	print(path)
 	#Eric, change this so it's specific to your directory. Be sure that you are going all the way down to \static\texts\\
 	#path = r"C:\Users\Eric\Pykov-Text-Generator\static\texts\\"
 	for f in os.listdir(path):
-		text_file = open(os.path.join("static", "texts", f), 'r')
+		text_file = open(os.path.join("Pykov", "static", "texts", f), 'r')
 		text = text_file.read()
 		default_texts.append((re.split("\.", f)[0], text, pickle.dumps(Markov.gen_relation(text)), 1))
 	for text in default_texts:
