+++ Pykov/get_dialog.py
@@ -1,0 +1,8 @@
+import re
+
+f = open("static/texts/Pulp Fiction.txt", 'r')
+
+for line in f:
+	val = re.match("\s{25}(?!\s)(\w+ )*", line)
+	if val != None:
+		print(val.string)
