+++ Pykov/templates/index.html
@@ -30,8 +30,8 @@
 					</form>
 				</div>
 			{% else %}
-				<a class='alignright button' href="login.html">Login</a>
-				<a class='alignright button' href="signup.html">Sign Up</a>
+				<a class='alignright button' href="login">Login</a>
+				<a class='alignright button' href="signup">Sign Up</a>
 			{% endif %}
 		</div>
 		<div id=alltext>
