+++ Pykov/templates/login.html
@@ -16,7 +16,7 @@
   {% if error %}
 		<p class="error"><strong>Error:</strong> {{ error }}
 		{% endif %}
-		<form method="POST" action="/login.html">
+		<form method="POST" action="/login">
     <label><b>Username</b></label>
     <input type="text" placeholder="Enter Username" name="userName" required>
 
@@ -25,7 +25,7 @@
         
     <button type="submit">Login</button>
 	</form>
-	<a href="signup.html"><button type="submit">Sign Up</button></a>
+	<a href="signup"><button type="submit">Sign Up</button></a>
 
 </body>
 </html>
