+++ run.py
@@ -1,0 +1,2 @@
+from Pykov import app
+app.run(debug=True)
